Refactor tests
* Delete unnecessary tests
* Move some tests into 'systemTests' folder
