+++ src/test/java/com/mathworks/ci/systemTests/RunMatlabBuildIT.java
@@ -2,21 +2,31 @@ package com.mathworks.ci.systemTests;
 
 import com.mathworks.ci.MatlabBuildWrapperContent;
 import com.mathworks.ci.Message;
+import com.mathworks.ci.TestMessage;
 import com.mathworks.ci.UseMatlabVersionBuildWrapper;
 import com.mathworks.ci.freestyle.RunMatlabBuildBuilder;
 import com.mathworks.ci.freestyle.options.BuildOptions;
+import hudson.EnvVars;
 import hudson.model.FreeStyleBuild;
 import hudson.model.FreeStyleProject;
 import hudson.model.Result;
+import hudson.slaves.DumbSlave;
+import hudson.slaves.EnvironmentVariablesNodeProperty;
+import hudson.tasks.Builder;
 import org.htmlunit.html.HtmlAnchor;
 import org.htmlunit.html.HtmlElement;
 import org.htmlunit.html.HtmlPage;
+import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.job.WorkflowJob;
+import org.jenkinsci.plugins.workflow.job.WorkflowRun;
 import org.junit.*;
 import org.jvnet.hudson.test.ExtractResourceSCM;
 import org.jvnet.hudson.test.JenkinsRule;
 import org.xml.sax.SAXException;
 
+import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
 import static junit.framework.Assert.assertTrue;
 
@@ -50,6 +60,20 @@ public class RunMatlabBuildIT {
         this.scriptBuilder = null;
     }
 
+    @Test
+    public void verifyBuildStepWithRunMatlab() throws Exception {
+        boolean found = false;
+        project.getBuildersList().add(scriptBuilder);
+        List<Builder> bl = project.getBuildersList();
+        for (Builder b : bl) {
+            if (b.getDescriptor().getDisplayName().equalsIgnoreCase(
+                    TestMessage.getValue("Builder.build.builder.display.name"))) {
+                found = true;
+            }
+        }
+        Assert.assertTrue("Build step does not contain Run MATLAB Build option", found);
+    }
+
     /*
      * Test to verify if Build FAILS when matlab command fails
      */
@@ -65,6 +89,7 @@ public class RunMatlabBuildIT {
         project.getBuildersList().add(tester);
         FreeStyleBuild build = project.scheduleBuild2(0).get();
         jenkins.assertBuildStatus(Result.FAILURE, build);
+        jenkins.assertLogContains(String.format(Message.getValue("matlab.execution.exception.prefix"), 1), build);
     }
 
 
@@ -82,6 +107,20 @@ public class RunMatlabBuildIT {
         FreeStyleBuild build = project.scheduleBuild2(0).get();
         jenkins.assertBuildStatus(Result.SUCCESS, build);
         jenkins.assertLogContains("buildtool check", build);
+    }
+    @Test
+    public void verifyDefaultTaskForNiTaskInput() throws Exception {
+        this.buildWrapper.setMatlabBuildWrapperContent(new MatlabBuildWrapperContent(Message.getValue("matlab.custom.location"), Utilities.getMatlabRoot()));
+        project.getBuildWrappersList().add(this.buildWrapper);
+        project.setScm(new ExtractResourceSCM(Utilities.getRunMATLABTestsData()));
+        RunMatlabBuildBuilder tester =
+                new RunMatlabBuildBuilder();
+        project.getBuildersList().add(tester);
+        FreeStyleBuild build = project.scheduleBuild2(0).get();
+        // Default test task fails
+        jenkins.assertBuildStatus(Result.FAILURE, build);
+        // Test task runs the test
+        jenkins.assertLogContains("testMultiply", build);
     }
 
     @Test
@@ -116,6 +155,49 @@ public class RunMatlabBuildIT {
         jenkins.assertLogNotContains("In check task", build);
         jenkins.assertLogContains("In dummy task", build);
     }
+
+    @Test
+    public void verifyPipelineOnSlave() throws Exception {
+        DumbSlave s = jenkins.createOnlineSlave();
+        String script ="node('!built-in') { runMATLABBuild() }";
+
+        WorkflowJob project = jenkins.createProject(WorkflowJob.class);
+        project.setDefinition(new CpsFlowDefinition(script, true));
+        WorkflowRun build = project.scheduleBuild2(0).get();
+
+        jenkins.assertLogNotContains("Running on Jenkins", build);
+    }
+
+    @Test
+    public void verifyMATLABscratchFileNotGenerated() throws Exception {
+        this.buildWrapper.setMatlabBuildWrapperContent(
+                new MatlabBuildWrapperContent(Message.getValue("matlab.custom.location"), Utilities.getMatlabRoot()));
+        project.getBuildWrappersList().add(this.buildWrapper);
+        scriptBuilder.setTasks("");
+        project.getBuildersList().add(this.scriptBuilder);
+        FreeStyleBuild build = project.scheduleBuild2(0).get();
+        File matlabRunner = new File(build.getWorkspace() + File.separator + "runMatlabTests.m");
+        Assert.assertFalse(matlabRunner.exists());
+    }
+
+    @Test
+    public void verifyBuildSupportsEnvVar() throws Exception {
+        EnvironmentVariablesNodeProperty prop = new EnvironmentVariablesNodeProperty();
+        EnvVars var = prop.getEnvVars();
+        var.put("TASKS", "compile");
+        var.put("BUILD_OPTIONS", "-continueOnFailure -skip test");
+        jenkins.jenkins.getGlobalNodeProperties().add(prop);
+        this.buildWrapper.setMatlabBuildWrapperContent(
+                new MatlabBuildWrapperContent(Message.getValue("matlab.custom.location"), Utilities.getMatlabRoot()));
+        project.getBuildWrappersList().add(this.buildWrapper);
+        scriptBuilder.setTasks("$TASKS");
+        scriptBuilder.setBuildOptions(new BuildOptions("$BUILD_OPTIONS"));
+        project.getBuildersList().add(scriptBuilder);
+        FreeStyleBuild build = project.scheduleBuild2(0).get();
+        jenkins.assertLogContains("compile", build);
+        jenkins.assertLogContains("-continueOnFailure -skip test", build);
+    }
+
 
     @Test
     public void verifyBuildSummaryInBuildStatusPage() throws Exception {
